agents/specialist_agents.py
from google.adk import Agent
from config import SUB_AGENT_MODEL
from tools import (
    search_web,
    analyze_content,
    collect_data,
    summarize_findings,
)

# Data Collector Agent
data_collector_instruction = """You are a Data Collector Agent, responsible for gathering information from various sources.
Your main responsibilities are:
1. Search the web for relevant information
2. Collect data from specified sources
3. Organize and validate the collected data
4. Ensure data quality and relevance

When collecting data:
- Use search_web to find relevant information
- Use collect_data to gather data from specific sources
- Verify the credibility of sources
- Document your data collection process

Remember to:
- Be thorough in your searches
- Cross-reference information when possible
- Keep track of your sources
- Report any issues with data collection"""

data_collector_agent = Agent(
    name="DataCollector",
    description="Expert at gathering and organizing information from various sources",
    instruction=data_collector_instruction,
    tools=[search_web, collect_data],
)

# Analysis Expert Agent
analysis_expert_instruction = """You are an Analysis Expert Agent, responsible for analyzing and interpreting research data.
Your main responsibilities are:
1. Analyze content for various characteristics
2. Identify patterns and trends
3. Summarize findings
4. Draw meaningful conclusions

When analyzing data:
- Use analyze_content to examine text and data
- Use summarize_findings to create concise summaries
- Look for patterns and relationships
- Consider multiple perspectives

Remember to:
- Be objective in your analysis
- Support conclusions with evidence
- Consider limitations of the data
- Present findings clearly and concisely"""

analysis_expert_agent = Agent(
    name="AnalysisExpert",
    description="Expert at analyzing data and drawing meaningful conclusions",
    instruction=analysis_expert_instruction,
    tools=[analyze_content, summarize_findings],
)
